+++ src/components/Tournament/Admin/TournamentAdminGroupsMatchs.vue
@@ -8,7 +8,7 @@ import {
 import { computed, ref } from 'vue';
 import FormField from '@/components/FormField.vue';
 import Modal from '@/components/Modal.vue';
-import AdminMatch from '@/components/Tournament/Admin/AdminMatch.vue';
+import MatchCard from '@/components/Tournament/MatchCard.vue';
 import { MatchTypeEnum } from '@/models/match';
 import type { TournamentDeref } from '@/models/tournament';
 import { useNotificationStore } from '@/stores/notification.store';
@@ -161,7 +161,7 @@ const launch_selected_matchs = async () => {
             class="border-b border-r"
           >
             <div class="flex flex-wrap justify-around p-1">
-              <AdminMatch
+              <MatchCard
                 v-for="match in matchs"
                 :key="match.id"
                 v-model="selected_matchs"
