+++ src/components/Tournament/SwissRoundTable.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
 import { computed } from 'vue';
-import KnockoutMatchCard from '@/components/Tournament/KnockoutMatchCard.vue';
+import MatchCard from '@/components/Tournament/MatchCard.vue';
+import { MatchTypeEnum } from '@/models/match';
 import type { SwissRound } from '@/models/swiss';
 import { groupBy } from '@/stores/tournament.store';
 
 const { swiss, teamPerMatch } = defineProps<{
-  // rounds: Record<string, Record<string, string | Record<string, string | number | boolean | undefined>[]>[]>;
   swiss: SwissRound;
   teamPerMatch: number;
 }>();
@@ -47,28 +47,21 @@ const roundCount = computed(() => 2 * swiss.min_score - 1);
         </div>
 
         <div
-          class="flex flex-col items-center gap-3"
+          class="flex flex-col items-center"
         >
           <div
             v-for="match in matchs"
             :key="match.id"
-            class="my-3 w-full divide-y px-2"
+            class="w-full"
           >
-            <KnockoutMatchCard
+            <MatchCard
               :match="match"
+              :match-type="{ type: MatchTypeEnum.SWISS, id: match.swiss }"
               :team-per-match="teamPerMatch"
             />
           </div>
         </div>
       </div>
     </div>
-    <!-- <div v-for="round_idx in roundCount" :key="round_idx" class="flex flex-col">
-      <div v-for="(match, idx) in rounds[round_idx]" :key="idx" class="my-3 divide-y">
-        <KnockoutMatchCard
-          :team-per-match="teamPerMatch"
-          :match="match"
-        />
-      </div>
-    </div> -->
   </div>
 </template>
